extracting object
